Custom hook for HomeScren
